Split the serviceworker store in modules, in order to allow the serviceworker.vue refactor
