+++ store/modules/serviceworker/index.ts
@@ -1,0 +1,5 @@
+export const name = 'serviceworker';
+export * from './serviceworker.state';
+export * from './serviceworker.actions';
+export * from './serviceworker.types';
+export * from './serviceworker.mutations';
