+++ store/modules/serviceworker/serviceworker.types.ts
@@ -1,0 +1,8 @@
+export const types = {
+    UPDATE_ARCHIVE: 'UPDATE_ARCHIVE',
+    UPDATE_SERVICEWORKER: 'UPDATE_SERVICEWORKER',
+    UPDATE_SERVICEWORKERPREVIEW: 'UPDATE_SERVICEWORKERPREVIEW',
+    UPDATE_WEBPREVIEW: 'UPDATE_WEBPREVIEW',
+    RESET_STATES: 'RESET_STATES',
+    UPDATE_SERVICEWORKERS: 'UPDATE_SERVICEWORKERS'
+};
